backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
scikit-learn>=1.3.0
numpy>=1.26.0
pandas>=2.1.0
groq>=0.15.0
python-dotenv>=1.0.0
